Fix most JS
