+++ app/assets/stylesheets/application.scss
@@ -19,3 +19,4 @@
  @import "bootstrap";
  @import "./global";
  @import "./bootstrap-litera.css";
+
