+++ app/javascript/controllers/application.js
@@ -6,4 +6,13 @@ const application = Application.start()
 application.debug = false
 window.Stimulus   = application
 
+console.log('Welcome to Password Pusher! ( ◑‿◑)ɔ┏🍟--🍔┑٩(^◡^ )')
+console.log(' --> 🏝 May all your pushes be read once, stored securely and expired quickly.')
+        
+const prefersDarkScheme = window.matchMedia("(prefers-color-scheme: dark)");
+
+if (prefersDarkScheme.matches) {
+    document.body.classList.add('dark-mode')
+}
+
 export { application }
