+++ config/importmap.rb
@@ -12,3 +12,4 @@ pin "omgopass", to: "https://ga.jspm.io/npm:omgopass@3.2.1/index.js"
 pin "omgopass/random.js", to: "https://ga.jspm.io/npm:omgopass@3.2.1/random.browser.js"
 pin "clipboard", to: "https://ga.jspm.io/npm:clipboard@2.0.11/dist/clipboard.js"
 pin "spoiler-alert"
+pin "@rails/activestorage", to: "https://ga.jspm.io/npm:@rails/activestorage@7.0.4/app/assets/javascripts/activestorage.esm.js"
